Fix a build error that I introduced a few revisions ago. Strange that clang lets it through.

+++ OgreMain/include/OgreSearchOps.h
@@ -65,7 +65,7 @@ struct _finddata_t
 #endif
 #define _A_SUBDIR 0x10  /* Subdirectory */
 
-long _findfirst(const char *pattern, struct _finddata_t *data);
+intptr_t _findfirst(const char *pattern, struct _finddata_t *data);
 int _findnext(intptr_t id, struct _finddata_t *data);
 int _findclose(intptr_t id);
